[Feat] 단건 조회시 조회수 상승 기능 추가 - recipe_view,recipe_like 를 int형 으로 바꾸고 기본값을 0,notnull 처리 - recipeRepository에 updateView 증가 쿼리 추가

+++ src/main/java/com/woowahan/recipe/domain/dto/recipeDto/RecipeFindResDto.java
@@ -16,8 +16,8 @@ public class RecipeFindResDto {
     private String recipe_title;
     private String recipe_body;
     private String userName;
-    private Long recipe_like;
-    private Long recipe_view;
+    private int recipe_like;
+    private int recipe_view;
     private LocalDateTime createdAt;
     private LocalDateTime lastModifiedAt;
 }
